Nest payload in a property named for the product
This avoids name clashes if different products are reporting telemetry to the same server.  For example suppose two different products A and B both use the same key "users" but in product A it represents is a number of users and in B it is a list of user names.  The server wishes to index telemetry messages in Elasticsearch, but Elasticsearch expects all field values with the same name to be of the same type, so without nesting it would not be possible to handle both products in one index.  Nesting under the product name means that "A.users" and "B.users" are treated as distinct fields.